Correct support for character encodings -- need to know source encoding, and explicitly convert from there to UTF-8 (instead of assuming it was 7-bit ASCII that I could simply pretend was UTF-8 aready, like before).
If the encoding isn't provided, assume windows-1252 if running
on Windows and latin-1 otherwise.

This is only implemented in the core and CLI and still needs to
be propagated to the GUI.

Also leave a couple comments on future changes needed to date
parsing for non-US locales and stability, and what's needed in
the GUI and self-contained-app wrappers to keep up with recent
CLI changes.