Use MacRoman as default encoding for Mac files
It seems I was always confused about the meaning of those 0xA6 bytes
used as the line separator; I'd been assuming Mac-format export files
were encoded in latin-1 and 0xA6 meant pipe which Palm had hijacked
for line end.

My new interpretation is that Mac-format export files are encoded in
MacRoman, where 0xA6 means paragraph separator (which is 0xB6 in
latin-1 and is visibly descended from 0xB6 in Unicode and UTF-8),
and this makes more sense.

Change the newline parsing to expect \xC2\xB6 (UTF-8 for that paragraph
character), and change the default encoding to MacRoman when running on
Macs.