Hack OAuth flow into GUI.
Remove username and password fields; synchronously authenticate
when Import is clicked.

A fair amount of refactoring/cleanup to get the details of the
notes export file (filename, encoding, and locale) out of the
Config object and into a new ExportFileDetails object. This
can be created on demand and passed around, instead of the
old semi-global approach.

As of this commit, the importer works about as well as before
(I tested both GUI and CLI) but feels somewhat hacky/ugly.

Improvements still to come for more responsive GUI feedback:
- asynchronous authentication flow with status and option to cancel
- automatically load and count notes as appropriate