OAuth decline cleanup.
Handle the case where user declines OAuth request (in browser;
still need better async model so they can cancel in our app).

Also changed handling of oauth_token parameter in http callback,
so instead of oauth_receiver.py accepting anything and returning
it to the client which asserts it matches what we expect, the client
provides the expected value to oauth_receiver.py which matches
against it. This is further narrows the set of http requests which
our little temporary webserver will respond to (in a way that's
probably wholly unnecessary).