Build improvements for Mac OS X.
Before, I was jumping through hoops to build with the Apple-provided
python2.5, which existed since OS X 10.5 and seemed to work on 10.5,
10.6 and 10.7. However, this required various contortions (sys.path
hacks to load wx) and I couldn't get the python2.5 provided with
10.8 to run on 10.5, meaning I can't package from 10.8 anyway.

I bit the bullet and did the following:
- build on 10.5 in a VM
- in that VM, install python.org Python 2.6.6, wxpython, py2app
- now I can use python 2.6, and don't have to worry about the
  32 vs 64 bit issue that otherwise comes up with wxpython 2.8

The resulting .app bundle has been tested and works on 10.5, 10.6
and 10.8.

It's even bigger than before, but bits are cheaper than my time.